Extract CF specific handler and stream multiplexer into separate file and struct
