+++ src/go.mod
@@ -1,23 +1,29 @@
-module github.com/cloudfoundry/pcap-release/src/pcap-api
+module github.com/cloudfoundry/pcap-release
 
-go 1.18
+go 1.19
 
 require (
 	code.cloudfoundry.org/bytefmt v0.0.0-20211005130812-5bb3c17173e5
 	code.cloudfoundry.org/cli v7.1.0+incompatible
+	github.com/containerd/go-runc v1.0.0
+	github.com/golang-jwt/jwt/v4 v4.4.2
 	github.com/google/gopacket v1.1.19
 	github.com/jessevdk/go-flags v1.5.0
 	github.com/onsi/ginkgo v1.16.4
 	github.com/onsi/gomega v1.16.0
-	github.com/sirupsen/logrus v1.8.1
+	github.com/sirupsen/logrus v1.9.0
+	github.com/vishvananda/netns v0.0.0-20220913150850-18c4f4234207
 	gopkg.in/yaml.v2 v2.4.0
 )
 
 require (
+	github.com/containerd/console v1.0.1 // indirect
 	github.com/fsnotify/fsnotify v1.4.9 // indirect
 	github.com/nxadm/tail v1.4.8 // indirect
+	github.com/opencontainers/runtime-spec v1.0.2 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
 	golang.org/x/net v0.0.0-20210428140749-89ef3d95e781 // indirect
-	golang.org/x/sys v0.0.0-20210423082822-04245dca01da // indirect
+	golang.org/x/sys v0.0.0-20220715151400-c0bba94af5f8 // indirect
 	golang.org/x/text v0.3.6 // indirect
 	gopkg.in/tomb.v1 v1.0.0-20141024135613-dd632973f1e7 // indirect
 )
